# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppObras API",
        "description": "API v1 de autenticação e conta",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base"
        }
    ],
    "paths": {
...
        "/api/v1/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Listar empresas do usuário",
                "description": "Retorna todas as empresas que o usuário autenticado pertence",
                "operationId": "44926c230232e0ab0f8674f0e7d1ef0c",
                "responses": {
                    "200": {
                        "description": "Lista de empresas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Construtora ABC"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Companies"
                ],
                "summary": "Criar empresa",
                "description": "Cria uma nova empresa e atribui o usuário como Admin",
                "operationId": "452e31b1f5c474b63614ab2a7138e467",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 2,
                                        "example": "Minha Construtora"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Minha Construtora"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/contractors": {
            "get": {
                "tags": [
                    "Contractors"
                ],
                "summary": "Listar empreiteiros",
                "description": "Retorna lista de empreiteiros da empresa",
                "operationId": "fc05a178905474e3e652ea3dd5114c46",
                "parameters": [
                    {
                        "name": "X-Company-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de empreiteiros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Construtora ABC"
                                                    },
                                                    "contact": {
                                                        "type": "string",
                                                        "example": "(11) 98765-4321"
                                                    },
                                                    "specialties": {
                                                        "type": "string",
                                                        "example": "Fundação, Estrutura"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contractors"
                ],
                "summary": "Criar empreiteiro",
                "description": "Cria um novo empreiteiro na empresa",
                "operationId": "5723ef720f05f0744659b666f05b91cc",
                "parameters": [
                    {
                        "name": "X-Company-Id",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Construtora ABC"
                                    },
                                    "contact": {
                                        "type": "string",
                                        "example": "(11) 98765-4321"
                                    },
                                    "specialties": {
                                        "type": "string",
                                        "example": "Fundação, Estrutura"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empreiteiro criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Construtora ABC"
                                                },
                                                "contact": {
                                                    "type": "string",
                                                    "example": "(11) 98765-4321"
                                                },
                                                "specialties": {
                                                    "type": "string",
                                                    "example": "Fundação, Estrutura"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}